Downloads/ai_development_mass_framework-20250622T063827Z-1-001/ai_development_mass_framework/requirements_complete.txt
# Universal MASS Framework - Consolidated Requirements
# ================================================
# Production-ready dependencies for the complete MASS Framework
# Includes all requirements for advanced data processing and AI capabilities
# Generated: June 22, 2025

# Core Web Framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
aiofiles>=23.1.0
websockets>=12.0

# Database & ORM
sqlalchemy>=2.0.0
alembic>=1.12.0

# Authentication & Security
PyJWT>=2.8.0
passlib[bcrypt]>=1.7.4
cryptography>=41.0.0
python-jose[cryptography]>=3.3.0

# HTTP & API
aiohttp>=3.9.0
requests>=2.31.0
httpx>=0.25.0

# Data Processing & Analytics
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
scipy>=1.11.0

# Machine Learning Extensions
joblib>=1.3.0
statsmodels>=0.14.0

# Data Validation & Serialization
pydantic>=2.0.0
marshmallow>=3.20.0

# Time Series and Statistics
python-dateutil>=2.8.0
pytz>=2023.3

# Async Processing
asyncio-mqtt>=0.13.0
aioredis>=2.0.0

# Configuration Management
python-dotenv>=1.0.0
pyyaml>=6.0.1
configparser>=5.3.0

# Logging & Monitoring
structlog>=23.1.0
prometheus-client>=0.19.0
psutil>=5.9.0

# Template Engine
jinja2>=3.1.0

# Development & Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0

# Optional: Production Deployment
gunicorn>=21.0.0
supervisor>=4.2.0

# Optional: Enhanced Data Processing
# Uncomment for advanced capabilities
# tensorflow>=2.13.0
# torch>=2.0.0
# transformers>=4.30.0
# spacy>=3.6.0

# Optional: Time Series Forecasting
# prophet>=1.1.4
# pystan>=3.5.0

# Utilities
typing-extensions>=4.7.0
click>=8.1.0
rich>=13.5.0
tqdm>=4.65.0

# File Processing
openpyxl>=3.1.0
xlsxwriter>=3.1.0

# JSON Processing
orjson>=3.9.0

# HTTP Client with better async support
httpcore>=0.18.0

# Background Task Processing
celery>=5.3.0
redis>=5.0.0

# Data Validation
cerberus>=1.3.4
jsonschema>=4.19.0

# Network & DNS
dnspython>=2.4.0

# Cryptographic utilities
bcrypt>=4.1.0
hashlib-compat>=1.0.0

# Email utilities (for notifications)
email-validator>=2.1.0

# URL parsing
yarl>=1.9.0

# Memory optimization
memory-profiler>=0.61.0

# Compression
lz4>=4.3.0
zstandard>=0.22.0

# Rate limiting
slowapi>=0.1.9

# Caching
cachetools>=5.3.0

# IP address utilities
ipaddress>=1.0.23

# UUID utilities
shortuuid>=1.0.11

# Data streaming
kafka-python>=2.0.2

# Optional: Geographic data processing
# geopy>=2.3.0
# shapely>=2.0.0

# Development only dependencies
# black>=23.7.0
# flake8>=6.0.0
# mypy>=1.5.0
# pre-commit>=3.3.0
